refactor(core): move expr_text/source_expr_text from fslc into fsl-core
Pure relocation, no behavior change (verified by the full existing test
suite, including rust/fslc/tests/cli_regression.rs's direct expr_text test
and every explain --readable / html / ledger caller). fslc re-exports both
functions unchanged so its 20+ existing call sites are untouched.

Needed so fsl-tools (which fslc depends on, not the other way around) can
reuse this exact expression-to-FSL-text renderer for the #326
controlled-language renderer's canonical-expression fallback, instead of
duplicating a ~150-line precedence-aware pretty-printer.
